requirements.txt
PySide6
pyqtgraph
numpy
pyserial
reportlab
PyQt6-WebEngine